+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # cargo-aprz
 
+## 0.14.0 - 2026-03-06
+
+### Fixed
+
+- Fix handling of features when walking transitive dependencies.
+
 ## 0.13.0 - 2026-03-01
 
 ### Added